Consolidate duplicate parse_array_type_syntax implementations
Extract the identical parse_array_type_syntax functions from
ConstraintGenerator and Sema into a single public function in
types.rs. This eliminates code duplication while maintaining
the same behavior.

Fixes rue-t94y
